Remove unnecessary OperatorFactories interface
