+++ core/trino-main/src/main/java/io/trino/operator/join/LookupJoinOperatorFactory.java
@@ -17,9 +17,9 @@ import com.google.common.collect.ImmutableList;
 import io.trino.operator.DriverContext;
 import io.trino.operator.HashGenerator;
 import io.trino.operator.InterpretedHashGenerator;
+import io.trino.operator.JoinOperatorType;
 import io.trino.operator.Operator;
 import io.trino.operator.OperatorContext;
-import io.trino.operator.OperatorFactories.JoinOperatorType;
 import io.trino.operator.OperatorFactory;
 import io.trino.operator.PrecomputedHashGenerator;
 import io.trino.operator.ProcessorContext;
