+++ core/trino-main/src/test/java/io/trino/operator/join/unspilled/JoinTestUtils.java
@@ -18,7 +18,6 @@ import io.airlift.units.DataSize;
 import io.trino.RowPagesBuilder;
 import io.trino.operator.Driver;
 import io.trino.operator.DriverContext;
-import io.trino.operator.OperatorFactories;
 import io.trino.operator.OperatorFactory;
 import io.trino.operator.PagesIndex;
 import io.trino.operator.PipelineContext;
@@ -55,7 +54,8 @@ import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static io.trino.operator.HashArraySizeSupplier.incrementalLoadFactorHashArraySizeSupplier;
-import static io.trino.operator.OperatorFactories.JoinOperatorType.innerJoin;
+import static io.trino.operator.JoinOperatorType.innerJoin;
+import static io.trino.operator.OperatorFactories.join;
 import static io.trino.sql.planner.SystemPartitioningHandle.FIXED_HASH_DISTRIBUTION;
 import static java.util.Objects.requireNonNull;
 
@@ -67,22 +67,20 @@ public final class JoinTestUtils
     private JoinTestUtils() {}
 
     public static OperatorFactory innerJoinOperatorFactory(
-            OperatorFactories operatorFactories,
             JoinBridgeManager<PartitionedLookupSourceFactory> lookupSourceFactoryManager,
             RowPagesBuilder probePages,
             boolean hasFilter)
     {
-        return innerJoinOperatorFactory(operatorFactories, lookupSourceFactoryManager, probePages, false, hasFilter);
+        return innerJoinOperatorFactory(lookupSourceFactoryManager, probePages, false, hasFilter);
     }
 
     public static OperatorFactory innerJoinOperatorFactory(
-            OperatorFactories operatorFactories,
             JoinBridgeManager<PartitionedLookupSourceFactory> lookupSourceFactoryManager,
             RowPagesBuilder probePages,
             boolean outputSingleMatch,
             boolean hasFilter)
     {
-        return operatorFactories.join(
+        return join(
                 innerJoin(outputSingleMatch, false),
                 0,
                 new PlanNodeId("test"),
